FLASK/Flask_API/requirements.txt
flask==3.0.2
python-dateutil==2.8.2
flask-cors